+++ test/ScopeSpec.js
@@ -293,6 +293,16 @@ describe('Scope', function(){
       root.$digest();
       expect(log).toEqual('12');
     });
+
+    it('should fire a $destroy event', function(){
+      var destructedScopes = [];
+      middle.$on('$destroy', function(event) {
+        destructedScopes.push(event.currentTarget);
+      });
+      middle.$destroy();
+      expect(destructedScopes).toEqual([middle]);
+    });
+
   });
 
 
@@ -395,5 +405,98 @@ describe('Scope', function(){
         expect($exceptionHandler).wasCalledWith(error);
       });
     });
+  });
+
+  describe('events', function(){
+    var log, child, grandChild, greatGrandChild;
+
+    beforeEach(function(){
+      log = '';
+      child = root.$new();
+      grandChild = child.$new();
+      greatGrandChild = child.$new();
+
+      root.id = 0;
+      child.id = 1;
+      grandChild.id = 2;
+      greatGrandChild.id = 3;
+
+      root.$on('myEvent', logger('<'));
+      root.$on('myEvent', logger('>'), true);
+      child.$on('myEvent', logger('<'));
+      child.$on('myEvent', logger('>'), true);
+      grandChild.$on('myEvent', logger('<'));
+      grandChild.$on('myEvent', logger('>'), true);
+      greatGrandChild.$on('myEvent', logger('<'));
+      greatGrandChild.$on('myEvent', logger('>'), true);
+    });
+
+    function logger(text) {
+      return function(event) {
+        log += event.currentTarget.id + text;
+      };
+    }
+
+    it('should fire event through capture and bubble phase', function(){
+      grandChild.$emit('myEvent');
+      expect(log).toEqual('0>1>2>2<1<0<');
+    });
+
+    it('should not propagate exceptions in bubble and capture', function(){
+      child.$on('myEvent', function(){ throw 'bubbleException'; });
+      child.$on('myEvent', function(){ throw 'captureException'; }, true);
+      grandChild.$emit('myEvent');
+      expect(log).toEqual('0>1>2>2<1<0<');
+      expect(mockHandler.errors[0]).toEqual('captureException');
+      expect(mockHandler.errors[1]).toEqual('bubbleException');
+    });
+
+    it('should allow cancalation of event propegation in bubble', function(){
+      child.$on('myEvent', function(event){ event.cancel(); });
+      grandChild.$emit('myEvent');
+      expect(log).toEqual('0>1>2>2<1<');
+    });
+
+    it('should allow cancalation of event propegation in capture', function(){
+      child.$on('myEvent', function(event){ event.cancel(); }, true);
+      grandChild.$emit('myEvent');
+      expect(log).toEqual('0>');
+    });
+
+    it('should remove event listener', function(){
+      function eventFn(){
+        log += 'abc;';
+      }
+
+      child.$on('abc', eventFn, true);
+      child.$emit('abc');
+      expect(log).toEqual('abc;');
+      log = '';
+      child.$removeOn('abc', eventFn, true);
+      child.$emit('abc');
+      expect(log).toEqual('');
+    });
+
+    it('should forward method arguments', function(){
+      child.$on('abc', function(event, arg1, arg2){
+        expect(arg1).toEqual('arg1');
+        expect(arg2).toEqual('arg2');
+      });
+      child.$emit('abc', 'arg1', 'arg2');
+    });
+
+    describe('event object', function(){
+      it('should have methods/properties', function(){
+        var event;
+        child.$on('myEvent', function(e){
+          expect(e.target).toBe(grandChild);
+          expect(e.currentTarget).toBe(child);
+          event = e;
+        });
+        grandChild.$emit('myEvent');
+        expect(event).toBeTruthy();
+      });
+    });
+
   });
 });